updates from bwcheng for gcn verification

+++ src/data_def.h
@@ -2,7 +2,7 @@
 
 struct cg_subtile1 {
     std::vector<int> mode_0;
-    std::vector<double> mode_vals;
+    std::vector<float> mode_vals;
 };
 
 struct cg_extents1 {
@@ -13,7 +13,7 @@ struct cg_extents1 {
 struct cg_subtile2 {
     std::vector<int> mode_0; 
     std::vector<int> mode_1;
-    std::vector<double> mode_vals;
+    std::vector<float> mode_vals;
 };
 
 struct cg_extents2 {
@@ -26,7 +26,7 @@ struct cg_subtile3 {
     std::vector<int> mode_0;
     std::vector<int> mode_1;
     std::vector<int> mode_2;
-    std::vector<double> mode_vals;
+    std::vector<float> mode_vals;
 };
 
 struct cg_extents3 {
@@ -41,7 +41,7 @@ struct tile1 {
     std::vector<int> crd1;
     std::vector<int> pos2;
     std::vector<int> crd2;
-    std::vector<double> vals;
+    std::vector<float> vals;
 };
 
 struct tile2 {
@@ -53,7 +53,7 @@ struct tile2 {
     std::vector<int> crd3;
     std::vector<int> pos4;
     std::vector<int> crd4;
-    std::vector<double> vals;
+    std::vector<float> vals;
 };
 
 struct tile3 {
@@ -69,13 +69,13 @@ struct tile3 {
     std::vector<int> crd5;
     std::vector<int> pos6;
     std::vector<int> crd6;
-    std::vector<double> vals;
+    std::vector<float> vals;
 };
 
 struct subtile1 { 
     std::vector<int> pos1; 
     std::vector<int> crd1;
-    std::vector<double> vals; 
+    std::vector<float> vals; 
 };
 
 struct subtile2 {
@@ -83,7 +83,7 @@ struct subtile2 {
     std::vector<int> crd1;
     std::vector<int> pos2;
     std::vector<int> crd2;
-    std::vector<double> vals;
+    std::vector<float> vals;
 };
 
 struct subtile3 {
@@ -93,5 +93,5 @@ struct subtile3 {
     std::vector<int> crd2;
     std::vector<int> pos3;
     std::vector<int> crd3;
-    std::vector<double> vals;
+    std::vector<float> vals;
 };
